backend/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
pandas
scikit-learn
redis